requirements.txt
torch
transformers
Pillow
tqdm
pandas
h5py
numpy